Separate actions into modules
